test: split boot and QR camera suites
