+++ tests/ui/boot/boot-decisions.test.tsx
@@ -1,43 +1,21 @@
+import "./boot-test-support"
 import { StrictMode } from "react"
-import { cleanup, render, screen, waitFor } from "@testing-library/react"
-import { afterEach, describe, expect, it, vi } from "vitest"
+import { render, screen, waitFor } from "@testing-library/react"
+import { describe, expect, it, vi } from "vitest"
 import {
-  BOOT_PROBE_TIMEOUT_MS,
   createBootController,
-  probeNetworkSentinel,
   readBootDecision,
-  readDeploymentVerdict,
   type BootDecisionSnapshot,
 } from "@/app/boot/boot-controller"
-import { WIPE_BROADCAST_CHANNEL } from "@/app/boot/boot-contract"
-import type { DeploymentVerdict } from "@/lib/deployment-headers"
 import { useBootState } from "@/app/boot/use-boot-state"
+import { recordReceipt, type ReceiptSubject } from "@/features/receipt-cache"
 import {
-  createWipeCoordinator,
-  installQuarantineBroadcastListener,
-  installWipeBroadcastListener,
-  QUARANTINE_REQUEST_TYPE,
-} from "@/app/boot/wipe-coordinator"
-import {
-  clearReceipts,
-  recordReceipt,
-  type ReceiptSubject,
-} from "@/features/receipt-cache"
-import * as databaseModule from "@/storage/database"
-import {
-  resetDatabaseAccessBarrierForTesting,
   SENSITIVE_WRITE_EXCLUSION_TIMEOUT_MS,
   withSensitiveWriteLock,
 } from "@/storage/database"
-import { decision, response, responseMissingHeader } from "../helpers/boot-fixtures"
-import { deferred } from "../helpers/deferred"
-import { MemoryStorage } from "../helpers/memory-storage"
-
-const localStorage = new MemoryStorage()
-Object.defineProperty(window, "localStorage", {
-  configurable: true,
-  value: localStorage,
-})
+import { decision, response } from "../../helpers/boot-fixtures"
+import { deferred } from "../../helpers/deferred"
+import { localStorage } from "./boot-test-support"
 
 interface FakeBootDatabaseOptions {
   countFailure?: "keys" | "pqIdentities"
@@ -100,148 +78,6 @@ function fakeBootDatabase(options: FakeBootDatabaseOptions = {}) {
     transaction,
   }
 }
-
-afterEach(() => {
-  clearReceipts()
-  cleanup()
-  window.localStorage.clear()
-  resetDatabaseAccessBarrierForTesting()
-  vi.useRealTimers()
-  vi.restoreAllMocks()
-  vi.unstubAllGlobals()
-})
-
-describe("destructive reachability probe", () => {
-  it("requires status 200 and an exact, untrimmed sentinel body", async () => {
-    const fetchImpl = vi.fn(async () => response("QR-CRYPT-REACHABLE"))
-    await expect(
-      probeNetworkSentinel({ fetchImpl, nonce: "fixed", timeoutMs: 50 }),
-    ).resolves.toMatchObject({ confirmed: true })
-    expect(fetchImpl).toHaveBeenCalledWith(
-      "/reachability-sentinel.txt?n=fixed",
-      expect.objectContaining({ method: "GET", cache: "no-store" }),
-    )
-
-    await expect(
-      probeNetworkSentinel({
-        fetchImpl: vi.fn(async () => response("QR-CRYPT-REACHABLE\n")),
-        timeoutMs: 50,
-      }),
-    ).resolves.toMatchObject({ confirmed: false })
-  })
-
-  it.each([
-    ["non-200", vi.fn(async () => response("QR-CRYPT-REACHABLE", 204))],
-    ["body mismatch", vi.fn(async () => response("captive portal"))],
-    ["fetch rejection", vi.fn(async () => Promise.reject(new TypeError("offline")))],
-  ])("treats %s as offline", async (name, fetchImpl) => {
-    const controller = createBootController({
-      fetchImpl,
-      readConnectivityHint: () => "offline",
-      readDecision: async () => decision(),
-    })
-    await controller.probe()
-    // A non-200 response is still an obtained response: evaluateDeploymentHeaders
-    // marks the status field failed, so the deployment gate latches blocked
-    // instead of confirming offline — and it is a checked failure, not a missing
-    // verdict. Body mismatch / fetch rejection stay offline-confirmed.
-    if (name === "non-200") {
-      expect(controller.getState()).toEqual({
-        kind: "blocked",
-        reason: "deployment-failed",
-      })
-      return
-    }
-    expect(controller.getState()).toEqual({ kind: "offline-confirmed" })
-  })
-
-  it("times out an unresponsive fetch", async () => {
-    vi.useFakeTimers()
-    const controller = createBootController({
-      fetchImpl: vi.fn(() => new Promise<Response>(() => undefined)),
-      probeTimeoutMs: BOOT_PROBE_TIMEOUT_MS,
-      readConnectivityHint: () => "offline",
-      readDecision: async () => decision(),
-    })
-    const pending = controller.probe()
-    await vi.advanceTimersByTimeAsync(BOOT_PROBE_TIMEOUT_MS)
-    await pending
-    expect(controller.getState()).toEqual({ kind: "offline-confirmed" })
-  })
-
-  it.each([
-    ["an offline sentinel", "not-the-sentinel"],
-    ["a confirming sentinel", "QR-CRYPT-REACHABLE"],
-  ])(
-    "keeps a user-requested reset failure terminal against %s still in flight",
-    async (_name, body) => {
-      const resolvers: Array<(value: Response) => void> = []
-      const controller = createBootController({
-        fetchImpl: () =>
-          new Promise<Response>((resolve) => {
-            resolvers.push(resolve)
-          }),
-        readDecision: async () => decision(),
-      })
-      const pending = controller.probe()
-
-      // The reset engaged the one-way barrier while that probe was still open.
-      controller.beginUserRequestedReset()
-      expect(controller.getState()).toEqual({ kind: "wiping" })
-      controller.reportResetFailure(["database"])
-
-      resolvers[0]?.(response(body))
-      await pending
-
-      expect(controller.getState()).toEqual({
-        kind: "partial-failure",
-        failedSteps: ["database"],
-      })
-    },
-  )
-
-  it("refuses to leave a destructive state through any later transition", async () => {
-    const controller = createBootController({
-      fetchImpl: async () => response("not-the-sentinel"),
-      readDecision: async () => decision(),
-    })
-    controller.reportResetFailure(["database", "database-verification"])
-    const terminal = {
-      kind: "partial-failure",
-      failedSteps: ["database", "database-verification"],
-    }
-
-    await controller.probe()
-    controller.start()
-    controller.stop()
-    expect(controller.nudgeDisplayOffline()).toBe(false)
-    expect(await controller.refreshRelayEligibility()).toBe(false)
-
-    expect(controller.getState()).toEqual(terminal)
-  })
-
-  it("ignores a stale probe generation", async () => {
-    const resolvers: Array<(value: Response) => void> = []
-    const fetchImpl = vi.fn(
-      () =>
-        new Promise<Response>((resolve) => {
-          resolvers.push(resolve)
-        }),
-    )
-    const controller = createBootController({
-      fetchImpl,
-      readConnectivityHint: () => "offline",
-      readDecision: async () => decision(),
-    })
-    const first = controller.probe()
-    const second = controller.probe()
-    resolvers[1]?.(response("not-the-sentinel"))
-    await second
-    resolvers[0]?.(response("QR-CRYPT-REACHABLE"))
-    await first
-    expect(controller.getState()).toEqual({ kind: "offline-confirmed" })
-  })
-})
 
 describe("boot decisions", () => {
   it.each([
@@ -947,405 +783,5 @@ describe("boot decisions", () => {
     await waitFor(() => expect(screen.getByText("wiped")).toBeInTheDocument())
     expect(fetchImpl).toHaveBeenCalledTimes(1)
     expect(performWipe).toHaveBeenCalledTimes(1)
-  })
-})
-
-describe("connectivity-hint gating (NS-02)", () => {
-  const failingFetch = () =>
-    vi.fn(async () => Promise.reject(new TypeError("offline")))
-  const succeedingSentinelFetch = () =>
-    vi.fn(async () => response("QR-CRYPT-REACHABLE"))
-  const hangingFetch = () => vi.fn(() => new Promise<Response>(() => undefined))
-
-  it("confirms offline when the sentinel fails and the hint is offline", async () => {
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "offline",
-      readDecision: async () => decision(),
-    })
-    await controller.probe()
-    expect(controller.getState()).toEqual({ kind: "offline-confirmed" })
-  })
-
-  it("blocks when the sentinel fails but the hint is online", async () => {
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "online",
-      readDecision: async () => decision(),
-    })
-    await controller.probe()
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "network-suspected",
-    })
-  })
-
-  it("blocks on an indeterminate hint", async () => {
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "indeterminate",
-      readDecision: async () => decision(),
-    })
-    await controller.probe()
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "network-suspected",
-    })
-  })
-
-  it("blocks instead of confirming offline via nudgeDisplayOffline", async () => {
-    const controller = createBootController({
-      fetchImpl: succeedingSentinelFetch(),
-      readConnectivityHint: () => "online",
-      readDecision: async () => decision(),
-    })
-    await controller.probe()
-    expect(controller.getState().kind).toBe("network-confirmed")
-
-    controller.nudgeDisplayOffline()
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "network-suspected",
-    })
-  })
-
-  it("blocks instead of confirming offline on an offline event while probing", async () => {
-    const target = new EventTarget()
-    const controller = createBootController({
-      fetchImpl: hangingFetch(),
-      readConnectivityHint: () => "online",
-      readDecision: async () => decision(),
-      eventTarget: target as Pick<Window, "addEventListener" | "removeEventListener">,
-    })
-    controller.start()
-    target.dispatchEvent(new Event("offline"))
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "network-suspected",
-    })
-  })
-
-  it("is terminal: stop, start, probe, and events leave it unchanged", async () => {
-    const target = new EventTarget()
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "online",
-      readDecision: async () => decision(),
-      eventTarget: target as Pick<Window, "addEventListener" | "removeEventListener">,
-    })
-    await controller.probe()
-    const blocked = controller.getState()
-
-    controller.stop()
-    expect(controller.getState()).toEqual(blocked)
-    controller.start()
-    expect(controller.getState()).toEqual(blocked)
-    await controller.probe()
-    expect(controller.getState()).toEqual(blocked)
-    target.dispatchEvent(new Event("online"))
-    target.dispatchEvent(new Event("offline"))
-    expect(controller.getState()).toEqual(blocked)
-    controller.release()
-    controller.acquire()
-    expect(controller.getState()).toEqual(blocked)
-  })
-
-  it("blocks without invoking the wipe executor", async () => {
-    // The state assertion is the one that fails if the lock is missing; a
-    // bare "performWipe was not called" assertion passes on today's code too,
-    // because a failing sentinel already reaches offline-confirmed without
-    // wiping. Keep both, and never keep only the second.
-    const performWipe = vi.fn()
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "online",
-      readDecision: async () => decision(),
-      performWipe,
-    })
-    await controller.probe()
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "network-suspected",
-    })
-    expect(performWipe).not.toHaveBeenCalled()
-  })
-
-  it("quarantines without wiping and without opening the database", async () => {
-    const performWipe = vi.fn()
-    // Spying on getDb is what actually proves the design's claim. Asserting that
-    // a later getDb() rejects only proves the barrier engaged — it would still
-    // pass if the lock path had opened a readwrite transaction first.
-    const openSpy = vi.spyOn(databaseModule, "getDb")
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "online",
-      readDecision: async () => decision(),
-      performWipe,
-    })
-    await controller.probe()
-
-    expect(controller.getState().kind).toBe("blocked")
-    expect(performWipe).not.toHaveBeenCalled()
-    expect(openSpy).not.toHaveBeenCalled()
-  })
-
-  it("enters blocked when a peer broadcasts a quarantine request", async () => {
-    const onQuarantine = vi.fn()
-    const stop = installQuarantineBroadcastListener({ onQuarantine })
-    const channel = new BroadcastChannel(WIPE_BROADCAST_CHANNEL)
-    channel.postMessage({ type: QUARANTINE_REQUEST_TYPE, version: 1 })
-    await vi.waitFor(() => expect(onQuarantine).toHaveBeenCalled())
-    stop()
-    channel.close()
-  })
-})
-
-describe("deployment verdict (NS-08)", () => {
-  const conformingSentinelFetch = () =>
-    vi.fn(async () => response("QR-CRYPT-REACHABLE"))
-  const failingFetch = () =>
-    vi.fn(async () => Promise.reject(new TypeError("offline")))
-  const sentinelFetchWithout = (header: string) =>
-    vi.fn(async () => responseMissingHeader("QR-CRYPT-REACHABLE", header))
-  const decisionWithVerdict = (verdict: DeploymentVerdict | undefined) => async () => {
-    if (verdict === undefined) {
-      const base = decision()
-      return {
-        wipeOnOnline: base.wipeOnOnline,
-        sensitiveDataExists: base.sensitiveDataExists,
-        cleanOrigin: base.cleanOrigin,
-        maintenanceTokenArmed: base.maintenanceTokenArmed,
-        resetChurnMb: base.resetChurnMb,
-        preferencesReadFailed: base.preferencesReadFailed,
-      }
-    }
-    return decision({ deploymentVerdict: verdict })
-  }
-
-  it("persists a passing verdict from the sentinel response", async () => {
-    const controller = createBootController({
-      fetchImpl: conformingSentinelFetch(),
-      readConnectivityHint: () => "offline",
-      readDecision: async () => decision(),
-    })
-    await controller.probe()
-    const stored = await readDeploymentVerdict()
-    expect(stored?.status).toBe("pass")
-  })
-
-  it("blocks when the sentinel response is missing security headers", async () => {
-    const controller = createBootController({
-      fetchImpl: sentinelFetchWithout("x-frame-options"),
-      readConnectivityHint: () => "offline",
-      readDecision: async () => decision(),
-    })
-    await controller.probe()
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "deployment-failed",
-    })
-  })
-
-  it("blocks on a later boot when no verdict was ever persisted", async () => {
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "offline",
-      readDecision: decisionWithVerdict(undefined),
-    })
-    await controller.probe()
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "deployment-unverified",
-    })
-  })
-
-  it("blocks on a later boot when the persisted verdict failed", async () => {
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "offline",
-      readDecision: decisionWithVerdict({
-        status: "fail",
-        failedFields: ["x-frame-options"],
-        checkedAt: 5,
-      }),
-    })
-    await controller.probe()
-    expect(controller.getState().kind).toBe("blocked")
-  })
-
-  it("honours this episode's pass without re-reading storage", async () => {
-    // network-confirmed -> PASS -> server stopped -> display offline
-    const controller = createBootController({
-      fetchImpl: conformingSentinelFetch(),
-      readConnectivityHint: () => "offline",
-      readDecision: async () => decision(),
-    })
-    await controller.probe()
-    expect(controller.getState().kind).toBe("network-confirmed")
-    controller.nudgeDisplayOffline()
-    expect(controller.getState()).toEqual({ kind: "offline-confirmed" })
-  })
-
-  it("blocks without wiping when the verdict fails", async () => {
-    const performWipe = vi.fn()
-    const controller = createBootController({
-      fetchImpl: sentinelFetchWithout("cross-origin-opener-policy"),
-      readConnectivityHint: () => "offline",
-      readDecision: async () => decision(),
-      performWipe,
-    })
-    await controller.probe()
-    // Both halves matter: the state assertion is what detects a missing
-    // implementation, the wipe assertion is what detects an over-reaction.
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "deployment-failed",
-    })
-    expect(performWipe).not.toHaveBeenCalled()
-  })
-
-  it("refuses the router when the boot decision could not be read", async () => {
-    // readBootDecision returns FALLBACK_DECISION on an open failure
-    // (boot-controller.ts), which carries no verdict. Under D8 an
-    // absent verdict refuses, and it must still not wipe.
-    const performWipe = vi.fn()
-    const controller = createBootController({
-      fetchImpl: failingFetch(),
-      readConnectivityHint: () => "offline",
-      readDecision: () => Promise.reject(new Error("storage open failed")),
-      performWipe,
-    })
-    await controller.probe()
-    expect(controller.getState()).toEqual({
-      kind: "blocked",
-      reason: "deployment-unverified",
-    })
-    expect(performWipe).not.toHaveBeenCalled()
-  })
-})
-
-describe("WipeCoordinator order", () => {
-  it("clears session receipts during the buffer-drop step", async () => {
-    const subject: ReceiptSubject = {
-      kind: "sym",
-      recipientKeyId: "wipe-recipient",
-      envelopeHash: "wipe-envelope",
-    }
-    expect(recordReceipt(subject, 100)).toEqual({ kind: "first-seen" })
-
-    const coordinator = createWipeCoordinator({
-      engageBarrier: () => undefined,
-      disposeCrypto: () => undefined,
-      coordinateTabs: () => undefined,
-      withExclusiveLock: (operation) => operation(),
-      bestEffortReset: async () => ({ ok: true, failedSteps: [] }),
-    })
-    await coordinator.wipe({
-      reason: "online-detected",
-      resetChurnMb: 0,
-      resetTransient: () => undefined,
-    })
-
-    expect(recordReceipt(subject, 200)).toEqual({ kind: "first-seen" })
-  })
-
-  it("runs the fail-closed sequence once in its frozen order", async () => {
-    const order: string[] = []
-    const coordinator = createWipeCoordinator({
-      engageBarrier: () => {
-        order.push("1-barrier")
-      },
-      disposeCrypto: () => {
-        order.push("2-worker")
-      },
-      dropVaultKeyCacheAndReceipts: () => {
-        order.push("2-vault-cache")
-      },
-      coordinateTabs: () => {
-        order.push("4-tabs")
-      },
-      withExclusiveLock: async <T,>(operation: () => Promise<T>) => {
-        order.push("4-lock")
-        return operation()
-      },
-      bestEffortReset: async () => {
-        order.push("5-7-reset")
-        return { ok: true, failedSteps: [] }
-      },
-    })
-
-    const report = await coordinator.wipe({
-      reason: "online-detected",
-      resetChurnMb: 0,
-      endSession: () => order.push("0-relay"),
-      resetTransient: () => order.push("3-transient"),
-    })
-    await coordinator.wipe({
-      reason: "online-detected",
-      resetChurnMb: 0,
-      endSession: () => order.push("unexpected-relay"),
-      resetTransient: () => order.push("unexpected"),
-    })
-
-    expect(report).toEqual({ ok: true, failedSteps: [] })
-    expect(order).toEqual([
-      "0-relay",
-      "1-barrier",
-      "2-worker",
-      "2-vault-cache",
-      "3-transient",
-      "4-tabs",
-      "4-lock",
-      "5-7-reset",
-    ])
-  })
-
-  it("stops a relay before the barrier on a peer wipe broadcast", async () => {
-    let messageHandler: ((event: MessageEvent<unknown>) => void) | undefined
-    class FakeBroadcastChannel {
-      addEventListener(type: string, listener: (event: MessageEvent<unknown>) => void) {
-        if (type === "message") messageHandler = listener
-      }
-      close() {}
-      postMessage() {}
-      removeEventListener() {
-        messageHandler = undefined
-      }
-    }
-    vi.stubGlobal("BroadcastChannel", FakeBroadcastChannel)
-    const order: string[] = []
-    const remove = installWipeBroadcastListener(
-      {
-        endSession: () => order.push("relay-stop"),
-        resetTransient: () => order.push("transient"),
-      },
-      {
-        closeDatabase: () => order.push("close-db"),
-        disposeCrypto: () => order.push("crypto"),
-        dropVaultKeyCacheAndReceipts: async () => {
-          order.push("vault")
-        },
-        engageBarrier: () => order.push("barrier"),
-      },
-    )
-
-    messageHandler?.(
-      new MessageEvent("message", {
-        data: { type: "qr-crypt-wipe-request", version: 1 },
-      }),
-    )
-    expect(order.indexOf("relay-stop")).toBe(0)
-    expect(order.indexOf("relay-stop")).toBeLessThan(order.indexOf("barrier"))
-    await waitFor(() =>
-      expect(order).toEqual([
-        "relay-stop",
-        "barrier",
-        "crypto",
-        "vault",
-        "transient",
-        "close-db",
-      ]),
-    )
-    remove()
   })
 })
